(PUP-7482) Use correct mime type with http file sources
When retrieving a file whose source parameter is an HTTP(S) URI, send
the "Accept" content-type as "application/octet-stream" instead of
"binary". Earlier versions of puppet used to use "binary" as the
content-type, though it wasn't correct and wasn't compatible with
third-party web servers.

Also set the headers if the method is called without a block.

This doesn't fix the general issue of why we have 2 code paths for
creating HTTP connections. That will be addressed later.