(PUP-7482) Accept all content types with http file sources
In commit 89ea46b610, we modified puppet to use proper mime types in the
HTTP Accept header instead of the Puppet::Network::Format name. However,
we missed the case when using a file resource with http/https source
parameter, because the corresponding file_metadata and file_content
termini create Net::HTTP connections directly.

This commit makes puppet send "Accept: */*" for http/https file sources.
This is preferred over application/octet-stream, because some webservers
like Apache and IIS will respond with HTTP 406 if we request
octet-stream, but the file extension maps to some other well-known MIME
type, e.g. image/gif.

We also send "Accept-Encoding" so the server can send us compressed
content, and we set "User-Agent" based on the puppet version, like we do
for other HTTP connections.

This commit changes behavior in two ways. First, we send the above
headers when making the HEAD request prior to sending the "real"
request. And we send the above headers when a block is not given.