(PUP-8195) Add ability to register data type with Ruby implementation
This commit adds `Puppet::DataTypes#create_type`. It allows the user
to create new data types in Ruby in a way similar to how a function can
be created using `Puppet::Functions#create_function`.

In addition to the new `Puppet::DataType` module, this commit also adds
a `RubyDataTypeInstantiator` and `DataTypePath` to the loader framework
and the new loadable `:datatype`.