(PUP-8247) Fix Locator last-offset-cache
This fixes a problem in the Locator as it attempted to cache the last
offset given as an argument when wanting the line for that offset.
The code was broken in that it used a local variable instead of the
member variable and there were therefore never any cache hits.