(PUP-8262) Revert changes which made non-existent SMF services return :absent instead of :stopped
Prior to Puppet 4.10.9, non-existent Solaris SMF services reported their
state as :stopped, and could be defined as :stopped in Puppet code, to
ensure that unwanted services were not running, even if they were not
installed. In Puppet 4.10.9, this behavior was changed to cause
non-existent Solaris SMF services to report a state of :absent. This
change was too intrusive for our LTS branch and has caused workflow and
upgrade problems for users.

This commit consolidates the following reverts:

    Revert "(PUP-6797) service should check for smf service first"
    This reverts commit 3d7a3c4c1c6697047ae356c8765928f68618bc69.

    Revert "(PUP-6860) puppet on solaris 11 should support exists? for
services"
    This reverts commit dd30a54c9ff5fba9f980147b4a7af80442e009ed.

    Revert "(PUP-6797) service should check for smf service first"
    This reverts commit 3d7a3c4c1c6697047ae356c8765928f68618bc69.

    Revert "(PUP-6797) Update SMF tests for nonexistent service"
    This reverts commit 7085af0b7e508d86caa9e30bf44bed931d65f0df.

    Revert "(maint) (PUP-6767) remove redundant smf acceptance testing"
    This reverts commit b539fe581f372514e28f4c7e095d81dc8b0f500c.

    Revert "(PUP-8167) SMF provider: do not try to stop :absent
services"
    This reverts commit 7183a352288867392c1c934270c494e7b1270ab0.