(docs) Reword description to clarify truncated text
If you read the docs page describing the user resource, the summary for
the `expiry` parameter says `The expiry date for this user. Must be
provided...`, which might lead a user to believe that the parameter is
required.

This rewords the description so that the truncated version is clearly
not implying a required parameter.

I think there's a high chance that PUP-7663 originated as the user
mistakingly believing that they had to provide a value. (it's also
possible that they've specified defaults in `login.defs` and are trying
to override it for a specific user account.)