(PUP-8195) Fix glitch causing data-typed plain objects to fail on []
The evaluator would fail to call the underlying method for an data type
with a '[]' function declared when the mapped type was aa plain ruby
object. This is because the evaluator tested for the PuppetObject module
which a plain ruby object doesn't include.

This commit ensures that the type is inferred using the `TypeCalculator`
which in turn will consult the `ImplementationRegistry` if needed.