(PUP-8195) Minor speed improvement and bug-fix in serialization
This commit ensures that the `parameter_info` generated by the
`PObjectType` is cached since it is called on from many places. A minor
speed improvement/bug fix is also made to the deserializer where the
reader no longer needs to obtain the type of instance that it reads.
This also fixes a problem with reading plain ruby objects that are
mapped to data types. Before this commit, the reader would assume that
the class always had a _pcore_type method.