(PUP-7861) Update git mirror instructions
Prior to this commit, we were referencing an incorrect environment
variable to specify if the user wanted to hit a specific git mirror. In
Beaker, the environment variable we look for is either SERVER or
PROJECT_NAME_SERVER (i.e., PUPPET_SERVER). This commit updates
acceptance/README.md to accurately reflect this.