(PUP-8269) Move handling of future reserved word 'plan' to validator
Before this commit, the parser would emit a `Model::ReservedWord` with
future == true for the word `plan` when --tasks was not used. This led
to problems in the evaluator because it was not equipped to handle
reserved words that were allowed.

This commit changes the parser so that it no longer has special handling
of the word 'plan' (unless --tasks is used) and the responsibility of
issuing a warnning that a future reserved word is used is moved to the
validator `Checker4_0`.

The commit also removes the 'application', 'produces' and 'consumes'
from the hash of `FUTURE_RESERVED_WORDS` since they are now keywords.