(PUP-8271) Handle the special name for the init plan
Currently the typed_name method of a SmartPath does not tread init
specially. This updates the name check when a plan is instantiated to
handle this. There remains a bug where the name reported by list_plans
and list_tasks in PAL include the init portion. It's possible the fix
for that will make this code unecessary.