(PUP-8271) Make loader discovery find 'init' for plans and tasks
Before this commit, the discovery mechanism would incorrectly consider a
file named plans/init.pp to denote a plan named `<module-name>::init`.
The same thing was true files under tasks with names starting with
'init'.

This commit ensures that the `SmartPath` used for Plans and Tasks is
aware of the 'init' semantics when producing the `TypedName` for the
actual Plan or Target, thus avoiding errors about a name mismatch when
the corresponding entity is loaded.