(PUP-8270) Skip range max when presenting types with max == default
Several data types have a string representation that includes a range
with min and max values. Such representations will always come last in
an array. When creating type with a range from parameters, it's standard
that when the 'max' parameter is omitted it gets the default (unbounded)
value. This commit ensures that a 'max' with default value also is
omitted when the type is represented as a string. E.g. an integer
created as `Integer[2]` will presented as 'Integer[2]' instead of
'Integer[2, default]'.