(PUP-8268) aio setup should use new install method
This commit updates the aio setup steps to install puppet and
puppetserver with new methods added to beaker. These install the
packages based on data gathered from a yaml file, rather than trying to
figure out what the path to the artifact or repo file is.

This commit does not attempt to make the same change for the EC2 install
(as it's dependent on external infrastructure and the new install
methods rely on internal-only infrastructure). Nor does this commit
attempt to update the puppetserver nightly install. Nightlies is going
to change soon, so it is not worth updating those methods currently.

This commit also updates beaker and beaker-hostgenerator versions. This
change depends on code available in these newer releases. It also adds
beaker-puppet to the Gemfile, as we should be explicit about this
dependency since we are using methods that are specific to
beaker-puppet.