(pup-8288) Report exceptions during transactions as failed status.
Without this patch, exceptions during a transaction that are not caught will not be marked as failed.

Users and tools that use exit status information are not catching these failures due to the incorrect information.

This patch creates a flag that must be positively changed at the end of the transaction flow to allow the transaction to be marked with success.