(MAINT) Remove Travis before_install step
Prior to this commit the Travis config included a `before_install`
step, added in https://tickets.puppetlabs.com/browse/PUP-7920,
which installed bundler at `1.15.4` to workaround
https://github.com/travis-ci/travis-ci/issues/8357, which is fixed.

We discovered during research for MODULES-6339 that Travis now keeps
their matrices updated with the latest compatible bundler install.

For all of the builds in this matrix Travis is currently using
bundler `1.16.0` and the builds pass successfully.

This maintenance commit removes the pinned-version install of
bundler and instead returns to relying on Travis to build and
maintain images with compatible versions. This also follows the
advice in the comment in the config file on this step.