(maint) Allow numeric hostnames in hiera_in_templates test
This test creates a hash of hostname -> directory, and stores it in
YAML for hiera to query. If the hostname happens to be purely numeric,
the key in the hash will be a number, rather than a string. Facter
always returns the `hostname` fact as a string. This causes the hash
lookup in the puppet code to fail, and the manifest to fail to
compile.

The fix is to prepend a bit of non-numeric junk to the hostname, so
that the key in hiera is always string.