(PUP-8259) Make loader discovery report errors instead of failing
Before this commit, a function, type, plan, or task declaration that
contained errors would cause the loader discovery to raise an exception.
This is now changed so that the discovery will either raise warnings or
collect errors in an `error_collector` array that can be passed as an
optional argument. If the `error_collector` is provided, errors will not
be logged but instead converted into `Puppet::DataType::Error` instances
and added to the collector.