(maint) Remove unnecessary binary json test
This test was attempting to verify that the json terminus would error
in the face of files on disk with binary names. It doesn't actually do
that, and is also broken on High Sierra with APFS

Since this is not a particular valuable thing to test, instead of
attempting to repair the test it is simply being removed.