(PUP-8285) Check if file exists using the configured bucket
Previously, if you used a non-default bucket directory, the backup
command would only backup the file if it didn't exist in the default
bucket. This was a result of not passing the `bucket_path` request
option in the call to `Puppet::FileBucket::File.indirection.head`.

The same change is not need when calling `save` because in that case we
have an instance of Puppet::FileBucket::File that we want to save, and
it is already configured with the correct bucket_path.