(PUP-8357) Correctly match files for tasks
Previously, the existing_paths method for the FileBased loader was using
naive string matching to determine which files were associated with a
particular task. This resulted in the wrong set of files being found for
a task, in the case where the name of one task was a prefix of the name
of an adjacent task. For example, task `foo` would match both `foo.py`
and `foobar.py`. The desired behavior is that filenames need to match
all the way up to the file extension, in order to be associated with a
particular task.

This commit changes the existing_paths method to compare basenames
exactly, excluding file extensions.