(PUP-8356) Base fips specific behaviors without using facts
We need to adjust some settings based on whether fips mode is enabled
or not for FIPS compliance. Till now the later was done using a
fact (PUP-8141).
However that created a dependency in puppetserver which cannot
load facts in certain configurations like dev testing and running
from sources.
This change implements fips mode detection using a local method
instead of facter.