(PUP-8297) Don't accept already connected socket
Webrick's start method yields an already connected socket, so it is not
necessary to call accept on the socket. In ruby versions prior to 2.3.6 and
2.4.3, the accept call was a noop.

In 2.3.6 and 2.4.3, it is an error, because ruby those versions use a
non-blocking accept to handle clients that make an TCP connection, but
are slow to negotiate the TLS handshake[1,2]. On these versions, puppet
master would report:

    SSL_accept returned=1 errno=0 state=unknown state: unexpected record

Remove the call to Socket#accept.

[1] https://bugs.ruby-lang.org/issues/14005
[2] https://github.com/ruby/ruby/commit/2e728d51e70ed3756ad760c687a08b8487b0112f