(PUP-8343) Make rich data arguments to resources come through
Before this, the evaluator transformed all arguments to resources such
that some, but not all "rich data" was transformed to strings (notably
Regexp objects).
This was bad because a class, or define could then not rely on type
checking - a given Regexp would end up being a String which is
surprising because an error is expected.

The change in this commit is to not use the most aggressive 3x converter
for resources, and instead deal with rich data types in the
serialization (where it belongs).

The Serializer needed an update when turning a Regexp into a String, and
the output produced by Regexp#inspect is what puppet expects.

No tests are changed in the commit.