(PUP-7349) Generated tidy resources should have the same noop-ness
If a tidy resource's noop metaparameter is true or false, but not nil,
then generate file resources with the same noop-ness, otherwise the
files can be deleted.

If it's nil, then it means noop wasn't specified on the tidy resource,
so don't try to propagate it. We have to special case nil, because
you can't set noop to nil.

If the noop metaparameter is not set, but the Puppet noop setting is,
then just let puppet handle it, as it applies the setting to all
resources, including generated ones.