(PUP-6274) Raise Error if included classname is undef or empty string
Before this, undef and empty string given to include would not lead to
an error because the Class[main] is known internally as '' which is also
what undef becomes in the 3.x API.

There is no legit case where the class '' should be allowed in the
puppet language.

This commit raises an error if include is given an empty string or
undef/nil.