(PUP-8225) Ensure there is a consistent precedence when determining a node's environment
Previously when initializing a node from a data hash it attmepted to set the
environment but was instead accidentally setting a local variable. In addition
to this, when merging in a node's facts, an attempt was made to restore the
enviroment parameter to the pre-merged state but this was failing to do so
because it was using an ||= also because the method it was using to look up the
enviroment was primarily keying off of the value that had been changed.

We now properly set the environment attribute instead of a local variable in
'initialize_from_hash' and save the exisitng state of the enviroment parameter
before merging facts, so we can directly restore the value instead of having
to guess.

Additionally, we now initialize the environment_name by falling back to an
environment parameter if one is provided in the data hash. We do this to
maintain the enviroment parameter being in the precedence of enviroment lookup,
however this does move it behind the top level enviroment key in the data hash,
which appears to be where it was intended to belong.

There were some compiler related tests that were manually setting the topscope
environment variable, but this is no longer necessary as we are now
consistantly setting the environment variable on a node whenever it has an
environment.