(PUP-5426) Consider an IniFile::SettingsLine to be "different" from another if their line numbers differ
Previously, when we were trying to add a new setting to a section, if the
last setting in the section was identical (same prefix, name, infix,
value, and suffix) to a setting in another section that came first in the
file, we'd get confused about which one we should be adding the new
setting after, and add it after the one that appeared first in the file.

We now take the line number of the SettingLine into consideration when
testing equality (==), which means that Array#index will no longer give us
the index of the first SettingLine with the same
prefix/name/infix/value/suffix combination as the one we're looking for,
and will give us the index of the one that is at the same line number.