(PUP-8378) Provide graceful errors for any md5 runtime use attempts when in fips mode
In FIPS mode attempts to use md5 at runtime need to be disallowed while providing a graceful
error. This can happen when md5 checksum is explicitly specified in a file resource within
a manifest. User may also attempt to set the configuration settings of digest_algorithm and
supported_checksum_types with FIPS disallowed values.
This is related to PUP-8141.