(maint) Integer cast doesn't work with new fact
After the change to use an external fact instead of system_uptime to
deal with a windows issue (PUP-8363), the format of the time string we
get back changed. Early in the day this string can begin with zero,
which causes an Integer cast to interpret it as an octal. We can use
`to_i` to avoid this.