(PUP-3321) Should be able to alter an empty setting without adding excess lines.
Previously, if you tried to add a value to an existing setting that had no
value, it would place it on the following line. There was a new line being
added in the infix instead of the suffix when the file is written, so even if
you had previously set the value to '', a newline would already be there, and
would cause a Parse Error, but still edit the file. So each time you did this,
it would add more new lines and error, never actually editing the setting value
on the same line as the setting name. This solution strips any existing
newlines on the setting before attempting to set the value.