(PUP-8267) Support file or BHG string
HOSTS, BEAKER_HOSTS, and TEST_TARGET now behave the same. It's either a path to
a file containing the beaker host config or it's a beaker-hostgenerator string.
In the latter case, generate a host config file. In both cases, ENV['HOSTS']
will refer to a file for the duration of the run.

Add ci:test:quick to the rake description