(PUP-8258) Use Catalog cache in noop mode, but do not write to it
When running apply & agent in noop mode, we do want to be able to
retrieve previously cached catalogs according to the "normal" cache
lookup rules, but we should never write the retrieved catalog back to
the cache.

Since ignore_cache only affects retrieving from the cache, we can't
use it to turn off saving to the cache. Also, we do want to be able
to look up from the cache so even if it did turn off cache saving, we
wouldn't want to use it.

Now, we can explicitly specify that we do not wish to save to the
cache_class for the terminus by specifying ignore_cache_save, and we
set that depending on the value of noop.