(PUP-8378) Raise when setting md5 digests in FIPS mode
Previously it was possible to set digest_algorithm to md5 or
supported_checksum_types to md5 or md5lite, and puppet would crash
sometime later on FIPS enabled systems, e.g. when acting as a filebucket
client.

This commit adds validation to ensure we reject non-FIPS compliant
algorithms.  Also renames default_digest_alg and default_checksum_types.

This commit preserves the behavior of the supported_checksum_types
setting which defaults to non-lite digests, but will accept a lite
digest if specified.