(PUP-8409) Enable config print to print settings in a conf file compatible format.
When showing a setting with config print, without this change values are returned formatted for internal use.
Most of this change is to fix issues with environment_timeout which is a ttl setting and has a special value of 'unlimited' which was being printed as 'Infinity'.
With this change, returned values are formatted like the conf file expects.