(PUP-8432) Remove gettext-setup as a runtime gem dependency
Puppet maintains gem dependency information in multiple places. This
commit drops gettext-setup, and adds fast_gettext and locale as runtime
dependencies.

The ext/project_data.yaml is the source of puppet gem dependencies when it is
packaged as a gem.

* Remove gettext-setup as that is only needed when generating the pot files
  containing strings to be localized.
* Add fast_gettext and locale since we directly depend on them. Constrain
  fast_gettext to ~> 1.1.2 as 1.2 requires ruby 2.1.
* Bump the minimum facter dependency to 2.0.1 for consistency. Note facter
  2.0.0 was yanked.

The Gemfile is the source of gem dependencies when running from source via
bundler, and it includes everything in .gemspec (below). So no changes are
required.

The .gemspec is the source of gem dependencies when other code (eg modules)
   have a local checkout of puppet that they depend on.

* Bump minimum fast_gettext dependency to match project_data.yaml. Note

* It already depended on locale ~> 2.1 so that didn't need to change.

This is a follow up to commit 53aa50aa.