(PUP-8255) Strip 'default: ' from version number
Recently, rubygems started adding text 'default: ' to default versions of gems
when a user has more than one version installed. Puppet was interpreting this
text as part of the gem and would try to install a gem with the name 'default:'
and the version number that followed. Stripping this text in the gem package
provider resolves the issue.