(PUP-8439) Change how Timespan and Timestamp interprets one parameter
Prior to this commit, a `Timespan[x]` and `Timestamp[x]` would be
interpreted as `Timespan[x,x]` and `Timestamp[x,x]` respectively. This
was inconsistent with `Float` and `Integer` whose interpretation of
one parameter is that the second parameter is unlimited.

This commit changes the `Timespan` and `Timestamp` data types so that
they intepret one single parameter the same way as `Float` and
`Integer`, i.e. `T[x] == T[x, <unlimited>]`.