(PUP-8442) Params should allow default of `false`
Prior to this commit, parameters that had a default of false set
wouldn't populate properly. This is because in ruby, both `nil` and
`false` are considered to be 'falsey'. In effect, `defaultto false` was
treated the same as having no default value specified. This is not
ideal, considering there are cases where we treat a param set to `false`
differently then we would if that param were set to `nil`.

This commit fixes the bit in the code that short circuits when default
is set to `false`. Parameters with a default of `false` will not return
nil anymore.

This commit also fixes up some test cases that had worked around this
expectation, either by checking for 'falsiness', or not checking the
default value at all.