(PUP-8448) Only revoke certificates in puppet cert clean if any have already been signed
Previously, we would pass an empty list of hosts to the
Puppet::SSL::CertificateAuthority::Interface when asking it to revoke the
certificates if all of the certificates we were asked to clean were
unsigned. This broke the CA's expectations, and would cause it to raise an
exception. We now only ask the CA to revoke the list of certificates if
there is at least one signed certificate in the list of certificates to
clean.

This also fixes a spec that was attempting to test for this exact
scenario, but wasn't quite stubbing/expecting the method calls on
Puppet::SSL::CertificateAuthority::Interface correctly to detect the
issue.