(PUP-8443) Populate topscope environment variable directly from node's environment
Rather than relying on node.parameters to populate the topscope
environment variable, we now populate this variable by looking directly at
the node's environment. This means that if a node terminus sets the
environment on a node, but then later sets the node's parameters to a new
hash (instead of merging the new parameters into the existing one), we
will now properly populate the topscope environment with a value that
matches the node's environment, instead of having it be nil.