(maint) Refuse to destroy all certificates without an explicit list
Before the clean/destroy subcommand of cert gained the ability to also
revoke certificates when running, we would refuse to act on `--all`. This
restores the restriction that `--all` is invalid when attempting to
destroy certificates, and if a user really wants to destroy _all_ of the
certificates, they need to list them explicitly.