Revert "Merge pull request #6603 from kris-bosland/fix/master/pup-2868"
This reverts commit e9bf54fd34387fc6bd9ddc81f97cc9d270e2aa8b, reversing
changes made to 81c7c762ebd4759d727791f718961147d7c64c57.

Reverting this because each time `puppet config set` is called without a
section, it creates a duplicate [main] section. This arises because the
default section in code is :main (symbol), but the default section read
from the inifile is 'main' (string).