(maint) Remove useless array in void context in SMF service provider
Turns out that the SMF provider was never actually running `svcs -l
<name>` to see if a service was enabled and was instead creating an array
that would immediately get thrown away as it was used in void context. We
now run the svcs command, and check that it didn't raise a
Puppet::ExecutionFailure due to exiting with anything other than a 0
return value.