(PUP-8487) Remove MethodHelper
The functions in MethodHelper provided a generic way to supply optional
arguments to a constructor and send them to setters. In practice this
obscures the signature of the constructor and makes it very difficult to
figure out what arguments are being passed around. It also generated a
lot of garbage in constructing the names of setter methods.

I added temporary tracing to track what options were being set this way
and converted them all to explicit code in the initialize methods. Any
method that was using this before should now raise an exception when it
receives an unknown option (rather than silently ignoring it causing a
harder to understand error later).

The original logic of methodhelper was to ignore options with a nil
value. I have tried to reproduce this which may make the logic seem more
verbose than necessary. I found that many setters cannot handle being
passed nil. In some cases I've set instance variables directly where I
determined it was safe.

In a few instances I've deleted the options hash entirely as I don't
think it was being used.

I deleted a test in file_serving/base_spec.rb that was passing the same
argument as both a positional argument and in the option hash. As far as
I can tell nothing but the test actually does this, but before this
change it was possible (and the option hash argument would take
precedence) and it will now raise an exception.

I also deleted a test in indirector/indirection_spec.rb that appears to
be asserting that if you call the constructor with invalid arguments and
initialize throws an exception that the half initialized object is still
stored in the global set of indirections. This seems like less than
ideal behavior and I have not preserved it.