(PUP-686) Return changed exit code with --noop and count(out_of_sync) resources > 0
This should address the following bug:
https://tickets.puppetlabs.com/browse/PUP-686

When puppet is called with --noop all resources are noop-ed except those
that have an explicit 'noop => false'. Let's simply add the changes
status bit (0x2) if --noop was used and there are out of sync resources.
Handle all other resources as previously, so that if some resources were
indeed executed we set the return code as usual.