(PUP-8471) Prevent TypeSet loader from being replaced
Before this commit, the TypeSet loader that provides internal access
to the types defined locally in the set was replaced by the loader
that is associated with the file containing the type set source. This
replacement happened in the TypeParser when resolving the types
contained in the TypeSet. This commit prevents that replacement from
taking place.