(PUP-8470) Set :failonfail and :combine for consistency
Prior to this commit, we had modified a few calls to `execute` to pass
in a custom environment if one was set. This was great, except it had
some unintended consequences. The way `execute` is written, :failonfail
and :combine default to true unless any options are passed in. If
:failonfail or :combine are included in the options hash passed in,
that's fine, but otherwise, they default to false. This non-obvious
behavior means that any failures that occur as a result of running
execute are not surfaced. The puppet run will continue as if nothing bad
had happened at all. This commit makes sure we fail if we encounter any
failures when executing a command.