(PUP-8470) Check for dups if modifying gid or uid
Prior to this commit, when modifying a resource, we would always check
for any duplication. This was a change from previous behavior where, on
modification, we only checked for dups if changing the uid or the gid.
This commit updates modifycmd to be back in line with that behavior. The
resource we are modifying will always already exist unless we're
changing the uid or the gid. We don't want to fail in that case.

There are currently no tests for the gid behavior as the group type
doesn't really manage anything other than gid. Once groupadd.rb has been
modified to have support for `manages_members`, we'll be able to ensure
that `check_allow_dup` does not get called when managing group
membership.