(maint) Turn on Lint/UselessAccessModifier Rubocop check, and remove violations
The Lint/UselessAccessModifier can let us know when we think we've marked
something as public/protected/private, but haven't. Some examples of
common reasons for things being not-actually private are Ruby syntax, or
Ruby not actually supporting the thing in question being private at all.

We have a large number of violations of the Lint/UselessAccessModifier cop
for a few different reasons:
  * Trying to mark constants private. Ruby doesn't do this.
  * Trying to mark classes private. Ruby doesn't do this.
  * Trying to mark class methods defined using 'def self.foo' private.
    This doesn't work by placing the definition below 'private', and needs
    to be done by calling 'private_class_method :foo' after the method
    definition.
  * Marking the following methods as public when the methods were already
    in a public context.
  * Trying to mark class methods as protected. Ruby doesn't do this, as
    it's effectively the same as a private class method.

In all cases where we were trying to mark a method as private, but not
actually doing so, the method is now marked as private as long as it
didn't break any tests.

In cases where we were redundantly marking things as public, or where
making it private would have broken tests, the method is left as public.

In cases where we were trying to mark things as private and Ruby doesn't
support having them as private, the decoration has been removed.