(PUP-8501) Add MultiJson as a runtime dependency
In order to allow us to use the faster jackson backend for JSON parsing
server-side, we need to be able to choose which library to use for JSON
parsing based on what is available. MultiJson chooses from a
prioritized list at startup, based on which libraries are known to be the
fastest. On the server side, that will be Jackson (or rather, the JRuby
wrapper for Jackson, JrJackson), while on the agent side it is currently
the built-in JSON gem. Using MultiJson gives us options for
transparently switching to a faster backend agent-side in the future,
without requiring more code changes.