(maint) Allow reported times to accrue
Prior to this commit, if we were collecting time metrics under a given
label, and collected additional time metrics for that same label, the
original time metrics would be overwritten by the more recent times.
Because of this, we may have not been accurately reporting times. This
commit changes the `add_time` method to automatically accrue times for a
given label. However, we do still have a flag you can pass if you don't
actually want the sum of the times collected and want the last recorded
time to be the time presented in the report.