(PUP-7307) Only tidy unmanaged files
In the `tidy` type, adds a check to the `tidy?` filtering
function to ignore file paths that are already managed by
file resources in the catalog, since `tidy` cannot generate
file resources for those paths anyway.

In the unit test, generates a catalog when testing if files
should be tidied.