(PUP-8502) Use node terminus "plain" for lookup CLI unless --compile
This commit ensures that the lookup CLI uses the node terminus "plain"
unless given the --compile flag. The reason for this is that an ENC
can pull in classes and cause a compile anyway.

Acceptance tests must be created to ensure that the lookup CLI works
correctly in combination with an ENC. We need tests asserting that:

1. node terminus "plain" is used by the lookup CLI even when another
   terminus such as "exec" is configured.
2. a configured node terminus which is different from "plain" is used
   when the lookup CLI is passed the --compile flag.

Both cases have been tested manually.