(PUP-8501) Add wrapper to handle MultiJson fallback
This commit adds a utility wrapper that handles loading the built-in
JSON gem when MultiJson is not present (e.g. when running inside
pre-5.3.0 versions of puppetserver). The wrapper will attempt to load
MultiJson and if that fails, load JSON. The `load` and `dump` methods
correctly handle either case in the same way that MultiJson would.