(maint) Ensure the total time is reported last
Prior to this commit, the total time metric was being added somewhere in
the middle of our reported time metrics. We did have a sort check to
make sure total was being presented last, but this wasn't succeeded. In
`raw_summary`, we're stringifying the metrics keys, so 'total' doesn't
match :total. This commit makes sure 'total' is correctly identified. It
also adds a check to make sure we are actually sorting the time metrics
so that total is presented at the end of the list.

I also had to force transaction time to be a value larger than zero.
Apparently on windows puppet prints a message so fast the time gets
recorded as zero! This results in that time metric getting thrown out.
Since that's the only time metric we have for this test, windows
subsequently doesn't have a total time metric. The stub added in this
commit makes sure we do have a valid time to work with.